+++ builder_pkgs/mockito/CHANGELOG.md
@@ -9,6 +9,9 @@
 * Require dart_style >= 2.3.7, so that the current Dart language version can be
   passed to `DartFormatter`.
 * Add topics to `pubspec.yaml`.
+* Fix a bug where typedef-aliases in type arguments were not correctly
+  resolved.
+* Fix a bug where record types were not correctly resolved.
 
 ## 5.4.4